pyproject.toml
[tool.poetry]
name = "generative-music"
version = "0.1.0"
description = ""
authors = ["atsushieee <atsushi.tabata1204@gmail.com>"]
readme = "README.md"
packages = [{include = "generative_music"}]

[tool.poetry.dependencies]
python = ">=3.9.0,<3.10.0"
tensorflow = ">=2.8.0,<2.9.0"
numpy = "1.23.5"
miditoolkit = "^0.1.16"
scipy = "^1.11.1"
matplotlib = "^3.7.2"
tqdm = "^4.65.0"
pytest-mock = "^3.11.1"
pyyaml = "^6.0.1"

[tool.poetry.group.dev.dependencies]
black = "^22.10.0"
flake8 = "^6.0.0"
isort = "^5.10.1"
mypy = "^1.0.1"
pydocstyle = "^6.3.0"
pyproject-flake8 = "^6.0.0"
pytest = "^7.2.1"
toml = "^0.10.2"
types-pyyaml = "^6.0.12.11"

[tool.flake8]
max-line-length = 88
extend-ignore = "E203, E501"

[tool.black]
line-length = 88

[tool.mypy]
ignore_missing_imports = "True"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
